+++ ChangeLog.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Version [1.0.0](https://github.com/cedx/html.net/compare/v0.10.0...v1.0.0)
-- Added the `New-InputElement`, `New-TemplateElement` and `New-TextareaElement` cmdlets.
-- Added the `input`, `template` and `textarea` aliases.
+- Added the `New-FormElement`, `New-IframeElement`, `New-InputElement`, `New-TemplateElement` and `New-TextareaElement` cmdlets.
+- Added the `form`, `iframe`, `input`, `template` and `textarea` aliases.
 
 ## Version [0.10.0](https://github.com/cedx/html.net/compare/v0.9.0...v0.10.0)
 - Added the `New-ButtonElement`, `New-QElement`, `New-SelectElement`, `New-SlotElement` and `New-SourceElement` cmdlets.